+++ CHANGES.md
@@ -14,7 +14,7 @@
 - Refactored codebase (e.g. CLI handling, upgrade mechanism) for simplicity.
 - Let user specify regression test call in `"test-ancient"` alias.
 - Added `:no-tests` to CLI.
-- Implemented recursive artifact checking.
+- Implemented recursive artifact checking and upgrading.
 - Added `:recursive` to CLI.
 - Using [ancient-clj](https://github.com/xsc/ancient-clj) 0.1.4.
 - Using [rewrite-clj](https://github.com/xsc/rewrite-clj) 0.3.0