+++ CHANGES.md
@@ -1,6 +1,6 @@
 # CHANGES: `lein-ancient`
 
-## 0.5.0
+## 0.5.1
 
 - Removed colon from tasks (e.g. `upgrade` instead of `:upgrade`).
 - Warnings when using invalid/deprecated options/tasks.
@@ -21,6 +21,10 @@
 - Using [ancient-clj](https://github.com/xsc/ancient-clj) 0.1.4.
 - Using [rewrite-clj](https://github.com/xsc/rewrite-clj) 0.3.1
 - Added `latest` task to print latest artifact vector.
+
+## 0.5.0
+
+- botched release. me stupid.
 
 ## 0.4.4